requirements.txt
pygame==2.6.1
Pillow
